test(ImageTitle): Add BV test cases for padding testing
In this commit, we are doing the following:
- Boundary values for the
padding when providing extremely large aswell as extremely low paddings.
These are more considered to be edge-cases since for now there is no
boundary set for valid padding sizes, this means that both
Integer.MAX_VALUE and Integer.MIN_VALUE is both valid right now.

-
Fixing the test suite overall